[FLINK-26717][fs][core] Move s3 common utils to flink-core
This closes #19141
