+++ flink-core/src/main/java/org/apache/flink/core/fs/RefCountedTmpFileCreator.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.fs.s3.common.utils;
+package org.apache.flink.core.fs;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.util.function.FunctionWithException;
