+++ flink-core/src/test/java/org/apache/flink/core/fs/RefCountedFileWithStreamTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.fs.s3.common.utils;
+package org.apache.flink.core.fs;
 
 import org.junit.Assert;
 import org.junit.Rule;
